+++ demos/color-picker/highcontrast/index.ts
@@ -1,0 +1,9 @@
+import { ColorPicker,  } from './../../../src/color-picker/color-picker';
+import { enableRipple } from '@syncfusion/ej2-base';
+
+enableRipple(true);
+
+/**
+ * Default ColorPicker sample
+ */
+let colorPicker: ColorPicker = new ColorPicker({ }, '#picker');