+++ demos/uploader/allowExtensions.ts
@@ -1,0 +1,14 @@
+/**
+ * Default Uploader sample
+ */
+import { Uploader, FileInfo, AsyncSettings } from '../../src/uploader/uploader';
+import { Event } from '@syncfusion/ej2-base';
+
+let uploadObj: Uploader = new Uploader({
+    asyncSettings: {
+        saveUrl: 'http://104.238.131.174:8984/api/uploadbox/Save',
+        removeUrl: 'http://104.238.131.174:8984/api/uploadbox/Remove'
+    },
+    allowedExtensions: '.zip, .docx'
+});
+uploadObj.appendTo('#fileupload')